requirements.txt
streamlit
pandas
numpy==1.24.4
scikit-learn==1.2.2
xgboost
Pillow
